Fixed tests to start IIS Express when required.

+++ src/TimeTracker.Specifications/Support/WebApplication/IISExpressWebServer.cs
@@ -1,22 +1,25 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
+using System.Net.Http;
 
 namespace TimeTracker.Specifications.Support.WebApplication
 {
     public class IISExpressWebServer : IWebServer
     {
+        private readonly string BaseUrl;
         private readonly string IISExpressExe;
+        private readonly string IndexPageTitle;
         private readonly string WebsiteName;
 
         private Process Process;
 
-        public IISExpressWebServer(string websiteName)
-            : this(GetIISExpressExe(), websiteName)
+        public IISExpressWebServer(string websiteName, string baseUrl, string indexPageTitle)
+            : this(GetIISExpressExe(), websiteName, baseUrl, indexPageTitle)
         {
         }
 
-        public IISExpressWebServer(string iisExpressExe, string websiteName)
+        public IISExpressWebServer(string iisExpressExe, string websiteName, string baseUrl, string indexPageTitle)
         {
             if (!File.Exists(iisExpressExe))
             {
@@ -25,6 +28,8 @@ namespace TimeTracker.Specifications.Support.WebApplication
 
             IISExpressExe = iisExpressExe;
             WebsiteName = websiteName;
+            BaseUrl = baseUrl;
+            IndexPageTitle = indexPageTitle;
         }
 
         public void StartIfNotRunning()
@@ -57,12 +62,30 @@ namespace TimeTracker.Specifications.Support.WebApplication
 
         private bool IsRunning()
         {
-            // todo: test via ping?
-            return Process != null;
+            using (var client = new HttpClient())
+            {
+                client.BaseAddress = new Uri(BaseUrl);
+
+                var response = client.GetAsync("/").Result;
+
+                if (!response.IsSuccessStatusCode)
+                {
+                    return false;
+                }
+
+                var content = response.Content.ReadAsStringAsync().Result;
+
+                return content.Contains(string.Format("<title>{0}</title>", IndexPageTitle));
+            }
         }
 
         private void Stop()
         {
+            if (Process == null)
+            {
+                return;
+            }
+
             if (!Process.HasExited)
             {
                 Process.Kill();
